Added subjective probabilities

+++ src/healthcare/configuration/utility_reader.cc
@@ -16,7 +16,7 @@ namespace configuration {
 
 using ::boost::property_tree::ptree;
 
-std::function<double(int, int, int, int)> MakeUtilityFunc(
+  std::function<double(int, int, int, float)> MakeUtilityFunc(
     std::string func_str, std::unordered_map<std::string, double> util_consts,
     int max_age, int max_shocks, int max_fitness) {
   std::string expression_str = func_str;
@@ -58,7 +58,7 @@ std::function<double(int, int, int, int)> MakeUtilityFunc(
     exit(1);
   }
   return [expression, &age, &shocks, &fitness, &joy](
-             int age_in, int shocks_in, int fitness_in, int joy_in) {
+             int age_in, int shocks_in, int fitness_in, float joy_in) {
     age = static_cast<double>(age_in);
     shocks = static_cast<double>(shocks_in);
     fitness = static_cast<double>(fitness_in);
@@ -67,7 +67,7 @@ std::function<double(int, int, int, int)> MakeUtilityFunc(
   };
 }
 
-std::function<double(int, int, int, int)> ReadUtility(ptree util_config,
+std::function<double(int, int, int, float)> ReadUtility(ptree util_config,
                                                       int max_age,
                                                       int max_shocks,
                                                       int max_fitness) {
